Move objects back to separate files
